+++ MOVEit.MIT/Public/Disconnect-MITServer.ps1
@@ -1,7 +1,10 @@
 function Disconnect-MITServer {
     <#
     .SYNOPSIS
-        Disconnect from a MOVEit Transfer server.
+        Disconnect from a MOVEit Transfer server
+    .LINK
+        Revokes API token
+        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/Auth_RevokeToken        
     #>
     [CmdletBinding()]
     param (