+++ MOVEit.MIT/Public/Get-MITFolderContent.ps1
@@ -2,6 +2,15 @@ function Get-MITFolderContent {
     <#
     .SYNOPSIS
         Get MOVEit Transfer Folder Content
+    .LINK
+        Get content of the folder
+        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/GETapi/v1/folders/{Id}/content?Page={Page}&PerPage={PerPage}&SortField={SortField}&SortDirection={SortDirection}&Name={Name}-1.0        
+    .LINK
+        Get list of subfolders in folder
+        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/GETapi/v1/folders/{Id}/subfolders?Page={Page}&PerPage={PerPage}&SortField={SortField}&SortDirection={SortDirection}&Name={Name}-1.0
+    .LINK
+        Get list of files in folder
+        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/GETapi/v1/folders/{Id}/files?Page={Page}&PerPage={PerPage}&SortField={SortField}&SortDirection={SortDirection}&Name={Name}&NewOnly={NewOnly}-1.0                
     #>
     [CmdletBinding(DefaultParameterSetName='Content')]
     param (