+++ MOVEit.MIT/Public/Get-MITGroupMember.ps1
@@ -2,6 +2,9 @@ function Get-MITGroupMember {
     <#
     .SYNOPSIS
         Get MOVEit Transfer Group Member(s)
+    .LINK
+        Get a list of the members of a group
+        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/GETapi/v1/groups/{Id}/members?Page={Page}&PerPage={PerPage}-1.0        
     #>
     [CmdletBinding()]
     param (