+++ MOVEit.MIT/Public/Read-MITFile.ps1
@@ -1,7 +1,10 @@
 function Read-MITFile {
     <#
-        .SYNOPSIS
+    .SYNOPSIS
         Read (download) a MOVEit Transfer file/attachment
+    .LINK
+        Download file
+        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/GETapi/v1/files/{Id}/download-1.0        
     #>
     [CmdletBinding()]
     param (