+++ MOVEit.MIT/Public/Set-MITFolder.ps1
@@ -2,6 +2,9 @@ function Set-MITFolder {
     <#
     .SYNOPSIS
         Change a MOVEit Transfer Folder
+    .LINK
+        Partial Folder update
+        https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/PATCHapi/v1/folders/{Id}-1.0        
     #>
     [CmdletBinding()]
     param (