+++ MOVEit.MIT/Public/Set-MITFolderNotification.ps1
@@ -3,6 +3,7 @@ function Set-MITFolderNotification {
     .SYNOPSIS
         Change a MOVEit Transfer Folder Notification Setting(s)
     .LINK
+        Change folder update notifications
         https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/PATCHapi/v1/folders/{Id}/notifications-1.0        
     #>
     [CmdletBinding()]